add class and use it in changelog gen
